Bind handler into component context
With simple trick we can bind handlers into component context.
In constructor call `this.myHandler = this.myHandler.bind(this)`

It is more clean and react dont have to force rebiding events for every
render call

There is one more approach to this which is called "class properties"
and currently is in experimental state.
Babel plugin
https://babeljs.io/docs/plugins/transform-class-properties/
Small article with handler example
http://www.ian-thomas.net/autobinding-react-and-es6-classes/